Update server.cpp
update checkDateFormat function getting day and month and year in string format instead of integer to be able to truly check the length because for instance if we save 02 as a number of months it saves in INT as 2 and then the length would be 1, not 2. so to save 02 with it's own length I change the format to string.